Remove AppKit and UIKit utilities
The AppKit/UIKit files are now in the repository at https://github.com/douglashill/DHAppUIKit.git

The aim is to have these utilities available in frameworks, but UI-related utilities should not be imported in non-UI code so multiple frameworks are required. For dependency management (such as with Carthage), it is easiest if each framework has its own Xcode project and repository.